Move all mobs-related classes to models.mobs package
