--- roguelike-game/src/main/java/models/MobKind.java
@@ -1,3 +1,0 @@
-package models;
-
-public class MobKind {}
