+++ roguelike-game/src/main/java/models/mobs/AggressiveMobBehavior.java
@@ -1,4 +1,6 @@
-package models;
+package models.mobs;
+
+import models.Room;
 
 public class AggressiveMobBehavior implements MobBehavior {
     @Override
