+++ roguelike-game/src/main/java/models/mobs/ConfusedMob.java
@@ -1,6 +1,7 @@
-package models;
+package models.mobs;
 
 import java.util.Random;
+import models.Room;
 
 public class ConfusedMob extends Mob {
     private Mob mob;
