+++ roguelike-game/src/main/java/models/mobs/DragonMob.java
@@ -1,4 +1,5 @@
-package models;
+package models.mobs;
+
 
 /**
  * Concrete descendant of StrongMob class that is used by fantasy factory.
