+++ roguelike-game/src/main/java/models/mobs/FantasyMobGeneratorFactory.java
@@ -1,4 +1,5 @@
-package models;
+package models.mobs;
+
 
 /**
  * This class represents a concrete factory that can create fantasy mobs.
