+++ roguelike-game/src/main/java/models/mobs/MobBehavior.java
@@ -1,4 +1,6 @@
-package models;
+package models.mobs;
+
+import models.Room;
 
 interface MobBehavior {
     /**
