+++ roguelike-game/src/main/java/models/mobs/PassiveMobBehavior.java
@@ -1,4 +1,6 @@
-package models;
+package models.mobs;
+
+import models.Room;
 
 public class PassiveMobBehavior implements MobBehavior {
     @Override
