+++ roguelike-game/src/main/java/models/mobs/Player.java
@@ -1,4 +1,5 @@
-package models;
+package models.mobs;
+
 
 enum Gender {
     MALE,
