+++ roguelike-game/src/main/java/models/mobs/SkeletonMob.java
@@ -1,4 +1,5 @@
-package models;
+package models.mobs;
+
 
 /**
  * Concrete descendant of WeakMob class that is used by fantasy factory.
