+++ roguelike-game/src/main/java/models/mobs/StrongMob.java
@@ -1,4 +1,5 @@
-package models;
+package models.mobs;
+
 
 public class StrongMob extends Mob {
     public StrongMob(int posX, int posY, int health, int strength) {
