--- roguelike-game/src/test/java/models/ConfusedMobTest.java
@@ -1,11 +1,0 @@
-package test.models;
-
-import java.util.Random;
-
-import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
-
-import models.ConfusedMob;
-
-
-class ConfusedMobTest {}
