--- roguelike-game/src/test/java/models/MobKindTest.java
@@ -1,5 +1,0 @@
-package test.models;
-
-
-
-class MobKindTest {}
