+++ roguelike-game/src/test/java/models/RoomTest.java
@@ -9,8 +9,8 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import models.GameObject;
 import models.Room;
-import models.Mob;
-import models.Player;
+import models.mobs.Mob;
+import models.mobs.Player;
 
 
 class RoomTest {
