+++ roguelike-game/src/test/java/models/mobs/ConfusedMobTest.java
@@ -1,0 +1,8 @@
+
+package test.models.mobs;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+
+
+class ConfusedMobTest {}
