+++ roguelike-game/src/test/java/repositories/ConfigurationTest.java
@@ -1,6 +1,5 @@
 package test.repositories;
 
-import java.util.Random;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.Assert.*;
